Merge craft3 branch into master and update plugin name and code
This commit unifies the plugin codebase for Craft CMS 3.x and 4.x by
merging the craft3 branch into the master branch. It also renames the
plugin throughout the project to match the name on the plugin store and
improves the plugin code quality and consistency by refactoring the
internal code and variables.